tests: Move selected hunk recalculation
Selected-hunk recalculation now has a data owner, and command tests import
command refresh outcomes through that owner.

The data tests still sit inside the hunk navigation test module, leaving
the test tree to suggest the old owner for the recalculation path.

This commit addresses that by moving recalculation data cases into
tests/data/test_selected_change.hunk_recalculation.py and trimming stale
imports from test_hunk_tracking.py.

The next commit will add an import-boundary guard so the module split
remains explicit.
